try to refactor EditTaskFragment; add dependency for WheelView
