+++ app/src/main/res/values/strings.xml
@@ -6,9 +6,4 @@
     <string name="toast_double_click_to_exit">double click to exit</string>
     <string name="edit_task_fragment_alert_dialog_title">Are you sure?</string>
     <string name="edit_task_fragment_alert_dialog_message">Quit without saving?</string>
-    <string name="edit_task_fragment_alert_dialog_calcel">Cancel</string>
-    <string name="edit_task_fragment_alert_dialog_save">Save</string>
-    <string name="alert_dialog_title_new_group">New Group</string>
-    <string name="alert_dialog_message_add_new_group">Add a New Group</string>
-    <string name="edit_task_add_new_group_please_input_new_group_name">Please input new group name</string>
 </resources>
