+++ sharedbaseclass/build.gradle
@@ -22,5 +22,6 @@ dependencies {
     compile fileTree(include: ['*.jar'], dir: 'libs')
     testCompile 'junit:junit:4.12'
     compile 'com.android.support:appcompat-v7:23.2.1'
+    compile 'com.lantouzi.wheelview:library:1.1.2'
     compile project(':sharedutils')
 }
